twitter-sentiment-analysis/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
nltk>=3.9
matplotlib>=3.9
seaborn>=0.13
wordcloud>=1.9
fastapi>=0.111
uvicorn>=0.30
streamlit>=1.35
joblib>=1.4
pyyaml>=6.0
